# openapi/content.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Platform Content API",
        "description": "API сервиса content платформы",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/admin/v1/projects/{project}/content/images/search": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "GET /api/admin/v1/projects/{project}/content/images/search",
                "operationId": "content_index_api_admin_v1_projects_project_content_images_search",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
}